First implementation of the new plugin system. Text insertion is now done by the plugin using an util method. Only Gmail plugin implemented so far.
