+++ src/content/js/plugins/uservoice.js
@@ -1,3 +1,4 @@
+// TODO check if uservoice plugin still works
 App.plugin('uservoice', (function () {
 
     var parseList = function (list) {
